Sputnik v0.2.0 Request

This module wraps a http client

Functions

start(url)

Returns the following informations from the given url to the pid:

page body

request status code

request url

headers

Parameters

url: the URL to fetch via HTTP client

start(url, pid)

Asyncronously returns the following informations from the given url to the pid:

page body

request status code

request url

headers

Parameters

url: the URL to fetch via HTTP client

pid: the pid which will receive the output