Sputnik v0.1.0 Sputnik

This is the main entrance for the Sputnik program.

Functions

main(args)

This function is the main entrance for the CLI and it is not meant to be used directly.

# inside the project folder
$ mix escript.build
$ ./sputnik "http://spawnfest.github.io" --query "a" --query "h1,h2,h3"

start(url, queries \\ [])

Crawls a url and prints out the report.

Parameters

url: String that represents the initial url to crawl

queries: List of valid CSS selectors a strings

Examples

iex> Sputnik.start("https://spawnfest.github.io", ["a", "h1,h2,h3"])